added es 2

+++ es.1 - triangoli/Point.java
@@ -1,9 +1,9 @@
 
 public class Point {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public Point(float x, float y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -13,28 +13,28 @@ public class Point {
         this.y = 0;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public static float distanceBetween(Point a, Point b) {
-        float x1 = a.getX();
-        float x2 = b.getX();
-        float y1 = a.getY();
-        float y2 = b.getY();
-        return (float) Math.sqrt(((Math.pow((x2-x1),2))+Math.pow((y2-y1),2)));
+    public static double distanceBetween(Point a, Point b) {
+        double x1 = a.getX();
+        double x2 = b.getX();
+        double y1 = a.getY();
+        double y2 = b.getY();
+        return Math.sqrt(((Math.pow((x2-x1),2))+Math.pow((y2-y1),2)));
     }
 
     public String toString() {
